pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-wallet-rescue-ops"
version = "0.1.0"
description = "Wallet Rescue Ops v2: a high-stakes OpenEnv incident-response environment."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "fastapi>=0.115.0",
    "pydantic>=2.7.0",
    "requests>=2.31.0",
    "uvicorn>=0.30.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "wallet_rescue_ops.server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["wallet_rescue_ops*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

